+++ Riot/ViewController/LanguagePickerViewController.m
@@ -109,7 +109,7 @@
     [super viewWillAppear:animated];
 
     // Screen tracking
-    [[AppDelegate theDelegate] trackScreen:@"CountryPicker"];
+    [[Analytics sharedInstance] trackScreen:@"CountryPicker"];
 }
 
 - (void)tableView:(UITableView *)tableView willDisplayCell:(UITableViewCell *)cell forRowAtIndexPath:(NSIndexPath *)indexPath;