+++ Riot/ViewController/RoomSearchViewController.m
@@ -101,7 +101,7 @@
     }
 
     // Screen tracking
-    [[AppDelegate theDelegate] trackScreen:@"RoomsSearch"];
+    [[Analytics sharedInstance] trackScreen:@"RoomsSearch"];
     
     // Enable the search field by default at the screen opening
     if (self.searchBarHidden)