+++ Riot/ViewController/SettingsViewController.m
@@ -423,7 +423,7 @@ typedef void (^blockSettingsViewController_onReadyToDestroy)();
     [super viewWillAppear:animated];
 
     // Screen tracking
-    [[AppDelegate theDelegate] trackScreen:@"Settings"];
+    [[Analytics sharedInstance] trackScreen:@"Settings"];
     
     // Release the potential pushed view controller
     [self releasePushedViewController];
@@ -2854,7 +2854,7 @@ typedef void (^blockSettingsViewController_onReadyToDestroy)();
         [[NSUserDefaults standardUserDefaults] setBool:NO forKey:@"enableCrashReport"];
         [[NSUserDefaults standardUserDefaults] synchronize];
         
-        [[AppDelegate theDelegate] stopAnalytics];
+        [[Analytics sharedInstance] stop];
         
         // Remove potential crash file.
         [MXLogger deleteCrashLog];
@@ -2865,7 +2865,7 @@ typedef void (^blockSettingsViewController_onReadyToDestroy)();
         [[NSUserDefaults standardUserDefaults] setBool:YES forKey:@"enableCrashReport"];
         [[NSUserDefaults standardUserDefaults] synchronize];
         
-        [[AppDelegate theDelegate] startAnalytics];
+        [[Analytics sharedInstance] start];
     }
 }